requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0